user authentication for a particular blog for crud operations have done

+++ routes/articles.js
@@ -2,21 +2,22 @@ const express = require('express');
 const Article = require('./../models/article')
 const router = express.Router();
 const path = require('path');
+const { requireAuth, currentUser, userName } = require('./../middleware/login_signup_mw');
 const app = express();
 
 
-const template_path = path.join(__dirname,'./templates/views');
+const template_path = path.join(__dirname, './templates/views');
 
 
-app.set('views' , template_path)
+app.set('views', template_path)
 
 
 router.get('/new', (req, res) => {
-    res.render('new', { article : new Article() })
+    res.render('new', { article: new Article() })
 })
 
 
-router.get('/edit/:id', async (req, res) => {   
+router.get('/edit/:id', async (req, res) => {
     const article = await Article.findById(req.params.id)
     res.render("edit", { article: article })
 })
@@ -33,13 +34,16 @@ router.get('/:slug', async (req, res) => {
 })
 
 
-router.post('/', async (req, res, next) => {
 
-    req.article =  new Article()
-    next()  
+router.post('/', currentUser, async (req, res, next) => {
+
+    // console.log(res.locals.user.email)   // current user is here
+
+    req.article = new Article()
+    next()
 }, PostAndPut('new'))
 
-router.put('/:id', async (req, res, next) => {
+router.put('/:id', currentUser, async (req, res, next) => {
     req.article = await Article.findById(req.params.id)
     next()
 }, PostAndPut('edit'))
@@ -52,8 +56,14 @@ function PostAndPut(path) {
         article.description = req.body.description
         article.markdown = req.body.markdown
 
+        // trying to save current user email in during post in articles collection
+        article.createdBy = res.locals.user.email;
+        // console.log(article.createdBy)
+
+
         try {
             article = await article.save()
+            // console.log(article)
             res.redirect(`/articles/${article.slug}`)
         }
         catch (error) {
@@ -62,9 +72,21 @@ function PostAndPut(path) {
     }
 }
 
-router.delete('/:id', async (req, res) => {
-    await Article.findByIdAndDelete(req.params.id)
-    res.redirect('/home')
+router.delete('/:id', currentUser, async (req, res) => {
+    const databaseUser = await Article.findById(req.params.id);
+    const UserMail = databaseUser.createdBy;
+    res.locals.UserMail = UserMail;
+    // console.log(databaseUser)
+    // console.log(res.locals.UserMail, res.locals.user.email)
+    const currentMail = res.locals.user.email;
+    // console.log(UserMail, currentMail);
+    if (UserMail === currentMail) {
+        await Article.findByIdAndDelete(req.params.id)
+        res.redirect('/home')
+    }
+    else{
+        res.send('You have not created this blog')
+    }
 })
